+++ tests/test_visitors/test_ast/test_complexity/test_function/test_elifs.py
@@ -8,7 +8,7 @@ from wemake_python_styleguide.visitors.ast.complexity.function import (
 )
 
 function_with_elifs = """
-{0}def test_module():
+def test_module():
     if 1 > 2:
         print(1)
     elif 2 > 3:
@@ -20,7 +20,7 @@ function_with_elifs = """
 """
 
 function_with_ifs = """
-{0}def test_module():
+def test_module():
     if True:
         print(1)
     if 2 > 3:
@@ -30,13 +30,13 @@ function_with_ifs = """
 """
 
 function_with_raw_if = """
-{0}def function():
+def function():
     if 1 == 2:
         print(1)
 """
 
 function_with_if_else = """
-{0}def function(param):
+def function(param):
     if param == 2:
         print(1)
     else:
@@ -44,7 +44,7 @@ function_with_if_else = """
 """
 
 function_with_ternary = """
-{0}def with_ternary(some_value):
+def with_ternary(some_value):
     return [some_value] if some_value > 1 else []
 """
 
@@ -56,15 +56,15 @@ function_with_ternary = """
     function_with_if_else,
     function_with_ternary,
 ])
-@pytest.mark.parametrize('mode', [
-    'async ',  # coroutine
-    '',  # regular function
-])
 def test_elif_correct_count(
-    assert_errors, parse_ast_tree, code, default_options, mode,
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+    mode,
 ):
     """Testing that all `if`/`elif`/`else` stuff is allowed."""
-    tree = parse_ast_tree(code.format(mode))
+    tree = parse_ast_tree(mode(code))
 
     visitor = FunctionComplexityVisitor(default_options, tree=tree)
     visitor.run()
@@ -75,18 +75,20 @@ def test_elif_correct_count(
 @pytest.mark.parametrize('code', [
     function_with_elifs,
 ])
-@pytest.mark.parametrize('mode', [
-    'async ',  # coroutine
-    '',  # regular function
-])
 def test_elif_incorrect_count(
-    assert_errors, parse_ast_tree, code, options, mode,
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    code,
+    options,
+    mode,
 ):
     """Testing that incorrect number of `elif` stuff is restricted."""
-    tree = parse_ast_tree(code.format(mode))
+    tree = parse_ast_tree(mode(code))
 
     option_values = options(max_elifs=1)
     visitor = FunctionComplexityVisitor(option_values, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [TooManyElifsViolation])
+    assert_error_text(visitor, '2')