+++ tests/test_visitors/test_ast/test_complexity/test_offset_visitor.py
@@ -43,36 +43,15 @@ def container():
 """
 
 nested_with = """
-if True:
+def container():
     with open('some') as temp:
         temp.read()
 """
 
-
 nested_while = """
-while True:
+def container():
     while True:
         continue
-"""
-
-
-nested_asyncfor = """
-async def container():
-    async for data in cursor:
-        pass
-"""
-
-
-nested_asyncwith = """
-async def container():
-    async with open('some') as temp:
-        pass
-"""
-
-nested_await = """
-async def container1():
-    async def container2():
-        await cursor
 """
 
 
@@ -84,13 +63,16 @@ async def container1():
     nested_try2,
     nested_with,
     nested_while,
-    nested_asyncfor,
-    nested_asyncwith,
-    nested_await,
 ])
-def test_nested_offset(assert_errors, parse_ast_tree, code, default_options):
+def test_nested_offset(
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+    mode,
+):
     """Testing that nested expression with default options works well."""
-    tree = parse_ast_tree(code)
+    tree = parse_ast_tree(mode(code))
 
     visitor = OffsetVisitor(default_options, tree=tree)
     visitor.run()
@@ -106,15 +88,17 @@ def test_nested_offset(assert_errors, parse_ast_tree, code, default_options):
     (nested_try2, 4),
     (nested_with, 1),
     (nested_while, 1),
-    (nested_asyncfor, 1),
-    (nested_asyncwith, 1),
-    (nested_await, 2),
 ])
 def test_nested_offset_errors(
-    assert_errors, parse_ast_tree, code, number_of_errors, options,
+    assert_errors,
+    parse_ast_tree,
+    code,
+    number_of_errors,
+    options,
+    mode,
 ):
     """Testing that nested expressions are restricted."""
-    tree = parse_ast_tree(code)
+    tree = parse_ast_tree(mode(code))
 
     option_values = options(max_offset_blocks=1)
     visitor = OffsetVisitor(option_values, tree=tree)
@@ -122,6 +106,32 @@ def test_nested_offset_errors(
 
     errors = [TooDeepNestingViolation for _ in range(number_of_errors)]
     assert_errors(visitor, errors)
+
+
+@pytest.mark.parametrize('code', [
+    nested_if,
+    nested_if2,
+    nested_for,
+    nested_with,
+    nested_while,
+])
+def test_nested_offset_error_text(
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    code,
+    options,
+    mode,
+):
+    """Testing that nested expressions are restricted."""
+    tree = parse_ast_tree(mode(code))
+
+    option_values = options(max_offset_blocks=1)
+    visitor = OffsetVisitor(option_values, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [TooDeepNestingViolation])
+    assert_error_text(visitor, '8')
 
 
 def test_regression282(assert_errors, parse_ast_tree, options):