+++ wemake_python_styleguide/violations/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 
-# TODO(@sobolevn): sort all new violations
 # TODO(@sobolevn): ensure that documentation is also sorted
 # TODO(@sobolevn): write tests for testing violation sorting order?