+++ wemake_python_styleguide/visitors/ast/complexity/counts.py
@@ -114,7 +114,7 @@ class MethodMembersVisitor(BaseNodeVisitor):
         for node, count in self._methods.items():
             if count > self.options.max_methods:
                 self.add_violation(
-                    TooManyMethodsViolation(node, text=node.name),
+                    TooManyMethodsViolation(node, text=str(count)),
                 )
 
     def visit_any_function(self, node: AnyFunctionDef) -> None: